+++ Base/src/main/scala/typeclass/Repeat.scala
@@ -85,7 +85,21 @@ trait BiRepeated[Expr[_], A, Z]
 	with ContraRepeated[Expr, A, Z]
 
 /** Predefined implicit implementations of Repeated */
-object Repeated extends LowPrioRepeated {
+object Repeated extends VersionSpecificRepeated with LowPrioRepeated {
+	private def apply[A, Acc, Z](
+		initFn: () => Acc,
+		appendFn: (Acc, A) => Acc,
+		resultFn: Acc => Z,
+	): Repeated[A, Z] = {
+		type Acc2 = Acc
+		new Repeated[A, Z] {
+			type Acc = Acc2
+			def init():Acc = initFn()
+			def append(acc:Acc, elem:A):Acc = appendFn(acc, elem)
+			def result(acc:Acc):Z = resultFn(acc)
+		}
+	}
+
 	/**
 	 * Repeated units results in a unit
 	 */
@@ -123,6 +137,18 @@ object Repeated extends LowPrioRepeated {
 			def result(acc:Acc):String = acc.toString
 		}
 		new RepeatedCodepoint()
+	}
+
+	/**
+	 * Creates a String consisting of the concatenation of the component strings
+	 * @since 0.1.1
+	 */
+	def idConcatenateString:Repeated[String, String] = {
+		Repeated.apply(
+			() => new StringBuilder,
+			(acc:StringBuilder, elem:String) => acc ++= elem,
+			(acc:StringBuilder) => acc.toString,
+		)
 	}
 }